+++ vdmj/src/main/java/com/fujitsu/vdmj/syntax/DefinitionReader.java
@@ -1749,8 +1749,8 @@ public class DefinitionReader extends SyntaxReader
 				}
 
 				LexIntegerToken lit = (LexIntegerToken)lastToken();
-				from = lit.value;
-				to = lit.value;
+				from = lit.value.longValue();
+				to = lit.value.longValue();
 
 				switch (nextToken().type)
 				{
@@ -1761,7 +1761,7 @@ public class DefinitionReader extends SyntaxReader
 						}
 
 						lit = (LexIntegerToken)readToken();
-						to = lit.value;
+						to = lit.value.longValue();
 						checkFor(Token.SET_CLOSE, 2265, "Expecting '{n1, n2}' after trace definition");
 						break;
